fix(flutter): resolve compilation errors in bulk upload and analytics screens
- bulk_upload_screen.dart: Add missing imports (dart:convert, BulkUploadState), fix unused variable, replace DesignTokens.lightText with textOnDark
- seller_analytics_screen.dart: Import OrderStatus from base_models, fix price-to-cents conversion using (item.price * 100).toInt()
- bulk_upload_viewmodel.dart: Import orignabaseProvider directly from orignabase_provider.dart, fix type casting for productId

Resolves critical analyzer errors preventing CI pipeline from passing.

+++ origna_gta/lib/features/products/bulk_upload_viewmodel.dart
@@ -2,7 +2,7 @@
 import 'dart:convert';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:origna_gta/core/providers.dart';
+import 'package:origna_gta/core/orignabase_provider.dart';
 import 'package:origna_gta/utils/csv_parser.dart';
 
 import 'bulk_upload_state.dart';
@@ -141,7 +141,7 @@ class BulkUploadViewModel extends StateNotifier<BulkUploadState> {
       final createdList = productIds.asMap().entries.map((e) {
         return CreatedProduct(
           index: e.key,
-          productId: e.value.toString(),
+          productId: (e.value as int).toString(),
           title: products[e.key]['title'].toString(),
         );
       }).toList();
